# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GophProfile API
  version: 1.0.0
  description: |
    API микросервиса управления аватарками пользователей.

    Сервис принимает изображения JPEG, PNG и WebP, сохраняет оригинал,
    асинхронно создаёт миниатюры 100x100 и 300x300 и позволяет получать
    аватарки, метаданные и историю загрузок.
servers:
  - url: http://localhost:9090
    description: Локальный запуск через Docker Compose

  - url: http://localhost:8080
    description: Локальный доступ через Kubernetes port-forward

  - url: http://gophprofile.local
    description: Локальный Kubernetes Ingress
...
paths:
  /api/v1/avatars:
    post:
      tags:
        - Avatars
      summary: Загрузить аватарку
      operationId: uploadAvatar
      security:
        - UserIDHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Изображение JPEG, PNG или WebP размером не более 10 MiB
      responses:
        "201":
          description: Аватарка принята и поставлена в очередь на обработку
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadAvatarResponse"

        "400":
          $ref: "#/components/responses/BadRequest"

        "413":
          $ref: "#/components/responses/PayloadTooLarge"

        "429":
          $ref: "#/components/responses/TooManyRequests"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    UserIDHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: Идентификатор пользователя, выполняющего защищённую операцию
...
  schemas:
    UploadAvatarResponse:
      type: object
      required:
        - id
        - user_id
        - url
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid

        user_id:
          type: string
          example: sergey

        url:
          type: string
          example: /api/v1/avatars/550e8400-e29b-41d4-a716-446655440000

        status:
          type: string
          enum:
            - pending
            - processing
            - completed
            - failed
          example: pending

        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Avatar not found

        details:
          type: string
          example: Additional error information

        max_size:
          type: integer
          format: int64
          minimum: 0
          example: 10485760
  responses:
    BadRequest:
      description: Некорректные параметры запроса или отсутствует X-User-ID
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

        text/plain:
          schema:
            type: string
          example: missing X-User-ID header
    TooManyRequests:
      description: Превышен допустимый лимит HTTP-запросов
      headers:
        Retry-After:
          description: Через сколько секунд рекомендуется повторить запрос
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: rate limit exceeded
    PayloadTooLarge:
      description: Размер файла превышает допустимый предел
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: File too large
            max_size: 10485760
...
    InternalServerError:
      description: Внутренняя ошибка сервиса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Internal server error
...